+++ R/LearnerClassifGlmnet.R
@@ -42,6 +42,7 @@
 #' autoplot(learner, type = "ggfortify")
 #' }
 autoplot.LearnerClassifGlmnet = function(object, type = "prediction", task = NULL, grid_points = 100L, expand_range = 0, theme = theme_minimal(), ...) { # nolint
+  assert_choice(type, choices = c("prediction", "ggfortify"), null.ok = FALSE)
   assert_has_model(object)
 
   switch(type,