+++ NEWS.md
@@ -1,5 +1,11 @@
 # mlr3viz (development version)
 
+- feat: Allow passing parameters to `precrec::autoplot()` (eg `show_cb`) when plotting `BenchmarkResult` and `ResampleResult` objects, using `type` = `roc` or `prc`.
+- refactor: Wrong `type` in `autoplot`s now gives hints of which ones to use.
+- refactor: Update `EnsembleFSResult.autoplot` to use the `active_measure` field.
+
+# mlr3viz 0.10.0
+
 - Add plot for `LearnerSurvCoxPH`.
 - Add plot for confidence intervals (`mlr3inferr`)